# .specify/specs/002-ai-character-chat/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChatZPT AI Character Chat API",
    "description": "AI 캐릭터와의 실시간 채팅을 위한 REST API",
    "version": "1.0.0",
    "contact": {
      "name": "ChatZPT Team",
      "email": "sendithere79@naver.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://api.chatzpt.com/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/characters": {
      "get": {
        "summary": "캐릭터 목록 조회",
        "description": "사용 가능한 AI 캐릭터 목록을 조회합니다.",
        "tags": [
          "Characters"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "캐릭터 이름 검색",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지당 결과 수",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "페이지 오프셋",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "캐릭터 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "characters": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Character"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "새 캐릭터 생성",
        "description": "새로운 AI 캐릭터를 생성합니다.",
        "tags": [
          "Characters"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "캐릭터 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "character": {
                      "$ref": "#/components/schemas/Character"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/user/preferences": {
      "get": {
        "summary": "사용자 설정 조회",
        "description": "사용자의 캐릭터별 설정을 조회합니다.",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "사용자 설정 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "preferences": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserPreference"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Character": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "char_123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "아리"
          },
          "description": {
            "type": "string",
            "example": "친근하고 활발한 AI 어시스턴트"
          },
          "personality": {
            "type": "string",
            "example": "친근하고 긍정적이며 도움이 되는 성격"
          },
          "avatar": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/avatar.jpg"
          },
          "systemPrompt": {
            "type": "string",
            "example": "당신은 아리라는 이름의 친근한 AI 어시스턴트입니다..."
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "example": "user_123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "personality",
          "systemPrompt",
          "isActive",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
      "CreateCharacterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "아리"
          },
          "description": {
            "type": "string",
            "example": "친근하고 활발한 AI 어시스턴트"
          },
          "personality": {
            "type": "string",
            "example": "친근하고 긍정적이며 도움이 되는 성격"
          },
          "avatar": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/avatar.jpg"
          },
          "systemPrompt": {
            "type": "string",
            "example": "당신은 아리라는 이름의 친근한 AI 어시스턴트입니다..."
          }
        },
        "required": [
          "name",
          "description",
          "personality",
          "systemPrompt"
        ]
      },
...
      "UserPreference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "pref_123e4567-e89b-12d3-a456-426614174000"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "user_123e4567-e89b-12d3-a456-426614174000"
          },
          "characterId": {
            "type": "string",
            "format": "uuid",
            "example": "char_123e4567-e89b-12d3-a456-426614174000"
          },
          "customPrompt": {
            "type": "string",
            "example": "항상 친근하게 대화해주세요"
          },
          "conversationStyle": {
            "type": "string",
            "enum": [
              "casual",
              "formal",
              "friendly"
            ],
            "example": "friendly"
          },
          "language": {
            "type": "string",
            "enum": [
              "ko",
              "en"
            ],
            "example": "ko"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "userId",
          "characterId",
          "conversationStyle",
          "language",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 100
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "offset": {
            "type": "integer",
            "example": 0
          },
          "hasMore": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "hasMore"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "요청이 올바르지 않습니다"
          },
          "details": {
            "type": "object",
            "example": {
              "field": "email",
              "reason": "INVALID_FORMAT"
            }
          }
        },
        "required": [
          "success",
          "error",
          "message"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "인증 실패",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "서버 내부 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
...
}